+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.5.0] - TBD
+## [3.5.0] - 2023-01-04
 ### Fixed
 - Throw exception if `cleanupDelay` can't be parsed instead of returning the default value.
 ### Changed